pyproject.toml
[project]
name = "anyrouter-check-in"
version = "0.1.0"
description = "Anyrouter or newapi system check in"
requires-python = ">=3.11"
dependencies = [
  "httpx[http2]>=0.24.0",
  "playwright>=1.40.0",
  "python-dotenv>=1.0.0"
]

[dependency-groups]
dev = [
  "ruff>=0.1.0",
  "pre-commit>=3.0.0"
]

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
select = ["ASYNC", "E", "F", "FAST", "I", "PLE"]
ignore = ["ASYNC109", "ASYNC230", "E101", "E402", "E501", "F841", "E731"]
unfixable = ["E101", "E402", "E501", "F841", "E731"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true
